refactor(app): extract PubTator genes controller (#346)
