+++ app/src/components/analyses/PubtatorNDDGenes.vue
@@ -327,14 +327,6 @@
 </template>
 
 <script setup lang="ts">
-import { ref, computed, watch, onMounted, onUnmounted } from 'vue';
-import { listPubtatorGenes } from '@/api/publication';
-
-// Import composables
-import { useToast, useUrlParsing, useTableData, useColumnTooltip } from '@/composables';
-import { useExcelExport } from '@/composables/useExcelExport';
-import { usePubtatorGenePublications } from '@/composables/usePubtatorGenePublications';
-
 // Small reusable components
 import TableSearchInput from '@/components/small/TableSearchInput.vue';
 import TablePaginationControls from '@/components/small/TablePaginationControls.vue';
@@ -344,11 +336,6 @@ import InlineHelpBadge from '@/components/small/InlineHelpBadge.vue';
 import AnalysisPanel from '@/components/analyses/AnalysisPanel.vue';
 import PubtatorPublicationDetail from '@/components/analyses/PubtatorPublicationDetail.vue';
 import {
-  applyPubtatorGenePrioritizationFilters,
-  createDefaultPubtatorGeneFilter,
-  type PubtatorGeneFilter,
-} from './pubtatorGeneFilters';
-import {
   formatCount,
   formatEnrichment,
   enrichmentVariant,
@@ -356,31 +343,8 @@ import {
   fdrStars,
   fdrClass,
   fdrTooltip,
-  createPubtatorGeneFields,
-  type PubtatorGeneFieldDefinition as FieldDefinition,
 } from './pubtatorEnrichmentDisplay';
-
-import { useUiStore } from '@/stores/ui';
-import { useRoute } from 'vue-router';
-
-// Types
-interface GeneItem {
-  gene_symbol: string;
-  gene_name: string;
-  gene_normalized_id?: string;
-  hgnc_id?: string;
-  publication_count: number;
-  oldest_pub_date?: string;
-  is_novel: number;
-  pmids?: string;
-  // Normalized enrichment metrics (issue #175); null until first refresh.
-  observed?: number | null;
-  background_count?: number | null;
-  enrichment_ratio?: number | null;
-  npmi?: number | null;
-  fisher_p?: number | null;
-  fdr_bh?: number | null;
-}
+import { usePubtatorGenesTable, type PubtatorGeneItem as GeneItem } from './usePubtatorGenesTable';
 
 // Props
 const props = withDefaults(
@@ -415,404 +379,46 @@ const emit = defineEmits<{
   'novel-count': [count: number];
 }>();
 
-// Composables
-const { makeToast } = useToast();
-const { getCompactTooltipText } = useColumnTooltip();
-const { filterObjToStr, filterStrToObj, sortStringToVariables } = useUrlParsing();
-const tableData = useTableData({
-  pageSizeInput: props.pageSizeInput,
-  sortInput: props.sortInput,
-  pageAfterInput: props.pageAfterInput,
-});
-const { isExporting, exportToExcel } = useExcelExport();
-
-const route = useRoute();
-
-// Destructure tableData
+// Table orchestration (filter/URL state, cursor-paginated load, enrichment
+// notice, page/sort handlers, export, fspec merge) lives in the composable so
+// this SFC stays a thin template shell.
 const {
   items,
   totalRows,
   perPage,
-  currentPage,
-  sortBy,
-  sort,
+  fields,
   loading,
   isBusy,
   downloading,
-  currentItemID,
-  prevItemID,
-  nextItemID,
-  lastItemID,
-  filter_string,
   pageOptions,
   removeFiltersButtonTitle,
   removeFiltersButtonVariant,
-} = tableData;
-
-// Table fields definition (includes issue #175 enrichment columns)
-const fields = ref<FieldDefinition[]>(createPubtatorGeneFields());
-
-// Component-specific filter
-const filter = ref<PubtatorGeneFilter>(createDefaultPubtatorGeneFilter());
-
-// Prioritization filter options
-const minPublications = ref<string>('all');
-const pubCountOptions = [
-  { value: 'all', text: 'All' },
-  { value: '2', text: '2+' },
-  { value: '5', text: '5+' },
-  { value: '10', text: '10+' },
-];
-
-const dateRange = ref<string>('all');
-const dateRangeOptions = [
-  { value: 'all', text: 'All time' },
-  { value: '1', text: 'Last 1 year' },
-  { value: '2', text: 'Last 2 years' },
-  { value: '5', text: 'Last 5 years' },
-];
-
-// Cursor pagination info
-const totalPages = ref(0);
-
-// Enrichment freshness notice (issue: surface a non-blocking notice when the
-// genes API reports the gene ranking is not yet computed / is stale). Defensive
-// — the meta fields may be absent (older API), in which case this stays null.
-const enrichmentNotice = ref<string | null>(null);
-
-// Per-gene publication-detail cache (extracted composable). Scoped to the
-// current filter/sort: resetCache() runs on every reload so an expanded row
-// never shows publications fetched under a previous query.
-const {
-  fetchPublications,
+  sortByArray,
+  enrichmentNotice,
+  minPublications,
+  pubCountOptions,
+  dateRange,
+  dateRangeOptions,
+  anyFilterContent,
+  getFilterContent,
+  setFilterContent,
+  applyPrioritizationFilters,
+  handlePageChange,
+  handlePerPageChange,
+  handleSortByUpdate,
+  filtered,
+  removeFilters,
+  removeSearch,
+  isExporting,
+  handleExcelExport,
+  copyLinkToClipboard,
+  handleRowExpand,
+  isLoadingPublications,
   getPublications,
-  isLoading: isLoadingPublications,
-  isCached,
-  resetCache: resetPublicationCache,
-  cancelAll: cancelAllPublicationFetches,
-} = usePubtatorGenePublications({ makeToast });
-
-// Computed: sortBy as properly typed array for BTable
-const sortByArray = computed(() => sortBy.value);
-
-// Computed: any filter content (string binding)
-const anyFilterContent = computed({
-  get: () => {
-    const content = filter.value.any.content;
-    return typeof content === 'string' ? content : '';
-  },
-  set: (value: string) => {
-    filter.value.any.content = value || null;
-  },
-});
-
-// Helpers for filter content binding
-const getFilterContent = (key: string): string => {
-  const content = filter.value[key]?.content;
-  return typeof content === 'string' ? content : '';
-};
-
-const setFilterContent = (key: string, value: string) => {
-  if (filter.value[key]) {
-    filter.value[key].content = value || null;
-    filtered();
-  }
-};
-
-// Computed: Novel gene count
-const novelCount = computed(
-  () => (items.value as GeneItem[]).filter((g) => g.is_novel === 1).length
-);
-
-// Watch novel count and emit to parent
-watch(
-  novelCount,
-  (count) => {
-    emit('novel-count', count);
-  },
-  { immediate: true }
-);
-
-// Helper: Parse PMIDs string to array
-const parsePmids = (pmids: string | undefined): string[] => {
-  if (!pmids) return [];
-  return pmids.split(',').filter((p) => p.trim() !== '');
-};
-
-// Helper: Truncate text
-const truncateText = (str: string | undefined, n: number): string => {
-  if (!str) return '';
-  return str.length > n ? `${str.slice(0, n)}...` : str;
-};
-
-// Handle row expansion - fetch publication data lazily via the cache composable
-const handleRowExpand = (item: GeneItem) => {
-  const pmids = parsePmids(item.pmids);
-  if (pmids.length > 0 && !isCached(item.gene_symbol)) {
-    fetchPublications(item.gene_symbol, pmids);
-  }
-};
-
-// Apply prioritization filters
-const applyPrioritizationFilters = () => {
-  filter.value = applyPubtatorGenePrioritizationFilters(filter.value, {
-    minPublications: minPublications.value,
-    dateRangeYears: dateRange.value,
-  });
-  currentItemID.value = 0;
-  filtered();
-};
-
-const loadData = async () => {
-  isBusy.value = true;
-
-  // The gene set is about to change (filter / sort / page). Drop the per-gene
-  // publication cache so an expanded row cannot show publications fetched under
-  // the previous query (correctness fix).
-  resetPublicationCache();
-
-  try {
-    const response = await listPubtatorGenes({
-      sort: sort.value,
-      filter: filter_string.value,
-      page_after: currentItemID.value,
-      page_size: String(perPage.value),
-      fields: props.fspecInput,
-    });
-
-    items.value = response.data || [];
-
-    if (response.meta && Array.isArray(response.meta) && response.meta.length > 0) {
-      const metaObj = response.meta[0] as {
-        totalItems?: number;
-        totalPages?: number;
-        prevItemID?: number | null;
-        currentItemID?: number;
-        nextItemID?: number | null;
-        lastItemID?: number | null;
-        currentPage?: number;
-        fspec?: FieldDefinition[];
-        // Optional enrichment-freshness fields from /pubtator/genes meta
-        // (camelCase, like executionTime/totalItems); absent on older responses.
-        enrichmentStatus?: string;
-        enrichmentRefreshedAt?: string;
-      };
-      totalRows.value = metaObj.totalItems || 0;
-      totalPages.value = metaObj.totalPages || 1;
-      prevItemID.value = metaObj.prevItemID || null;
-      currentItemID.value = metaObj.currentItemID || 0;
-      nextItemID.value = metaObj.nextItemID || null;
-      lastItemID.value = metaObj.lastItemID || null;
-
-      currentPage.value = metaObj.currentPage || 1;
-
-      if (metaObj.fspec && Array.isArray(metaObj.fspec)) {
-        fields.value = mergeFields(metaObj.fspec);
-      }
-
-      enrichmentNotice.value = deriveEnrichmentNotice(
-        metaObj.enrichmentStatus,
-        metaObj.enrichmentRefreshedAt
-      );
-    }
-
-    const uiStore = useUiStore();
-    uiStore.requestScrollbarUpdate();
-  } catch (error) {
-    makeToast(error, 'Error', 'danger');
-  } finally {
-    isBusy.value = false;
-  }
-};
-
-/**
- * Build a non-blocking notice string from the optional enrichment-freshness
- * meta. Returns null when the ranking is current or the fields are absent
- * (defensive: never throw on a shape the API may not yet send).
- */
-const deriveEnrichmentNotice = (
-  status: string | undefined,
-  refreshedAt: string | undefined
-): string | null => {
-  if (!status || status === 'current') return null;
-  if (status === 'missing') {
-    return 'Gene ranking not yet computed. Showing raw co-occurrence ordering until the first refresh.';
-  }
-  if (status === 'stale') {
-    const when = refreshedAt ? ` (last refreshed ${formatRefreshedAt(refreshedAt)})` : '';
-    return `Gene ranking may be out of date${when}.`;
-  }
-  return null;
-};
-
-/** Best-effort human-friendly date for the freshness notice. */
-const formatRefreshedAt = (value: string): string => {
-  const parsed = new Date(value);
-  return Number.isNaN(parsed.getTime()) ? value : parsed.toLocaleDateString();
-};
-
-// Handle page change
-const handlePageChange = (value: number) => {
-  if (value === 1) {
-    currentItemID.value = 0;
-  } else if (value === totalPages.value) {
-    currentItemID.value = lastItemID.value || 0;
-  } else if (value > currentPage.value) {
-    currentItemID.value = nextItemID.value || 0;
-  } else if (value < currentPage.value) {
-    currentItemID.value = prevItemID.value || 0;
-  }
-  filtered();
-};
-
-// Handle per page change
-const handlePerPageChange = (newSize: number | string) => {
-  perPage.value = parseInt(String(newSize), 10) || 10;
-  currentItemID.value = 0;
-  filtered();
-};
-
-// Rebuild filter string, reload data
-const filtered = () => {
-  const filterStringLoc = filterObjToStr(filter.value);
-  if (filterStringLoc !== filter_string.value) {
-    filter_string.value = filterStringLoc;
-  }
-  loadData();
-};
-
-// Clear all filters
-const removeFilters = () => {
-  filter.value = createDefaultPubtatorGeneFilter();
-  minPublications.value = 'all';
-  dateRange.value = 'all';
-  currentItemID.value = 0;
-  filtered();
-};
-
-// Clear the global "any" filter
-const removeSearch = () => {
-  filter.value.any.content = null;
-};
-
-// Handle sortBy updates from BTable
-const handleSortByUpdate = (newSortBy: Array<{ key: string; order: 'asc' | 'desc' }>) => {
-  sortBy.value = newSortBy;
-  currentItemID.value = 0;
-  // Extract sort string from array format for API
-  if (newSortBy && newSortBy.length > 0) {
-    const sortKey = newSortBy[0].key;
-    const sortDesc = newSortBy[0].order === 'desc';
-    sort.value = (sortDesc ? '-' : '+') + sortKey;
-  }
-  filtered();
-};
-
-// Copy link to clipboard
-const copyLinkToClipboard = () => {
-  const urlParam =
-    `sort=${sort.value}` +
-    `&filter=${filter_string.value}` +
-    `&page_after=${currentItemID.value}` +
-    `&page_size=${perPage.value}`;
-  const fullUrl = `${import.meta.env.VITE_URL}${route.path}?${urlParam}`;
-  navigator.clipboard.writeText(fullUrl);
-  makeToast('Link copied to clipboard', 'Info', 'info');
-};
-
-// Excel export handler
-const handleExcelExport = async () => {
-  if ((items.value as GeneItem[]).length === 0) {
-    makeToast('No data to export', 'Warning', 'warning');
-    return;
-  }
-
-  const exportData = (items.value as GeneItem[]).map((gene) => ({
-    gene_symbol: gene.gene_symbol,
-    gene_name: gene.gene_name,
-    publication_count: gene.publication_count,
-    background_count: gene.background_count ?? '',
-    enrichment_ratio: gene.enrichment_ratio ?? '',
-    npmi: gene.npmi ?? '',
-    fisher_p: gene.fisher_p ?? '',
-    fdr_bh: gene.fdr_bh ?? '',
-    oldest_pub_date: gene.oldest_pub_date || '',
-    source: gene.is_novel === 1 ? 'Literature Only' : 'Curated',
-    pmids: gene.pmids || '',
-  }));
-
-  try {
-    await exportToExcel(exportData, {
-      filename: `pubtator_genes_${new Date().toISOString().split('T')[0]}`,
-      sheetName: 'Gene Prioritization',
-      headers: {
-        gene_symbol: 'Gene Symbol',
-        gene_name: 'Gene Name',
-        publication_count: 'NDD Publication Count',
-        background_count: 'Background (Total) Publications',
-        enrichment_ratio: 'Enrichment Ratio',
-        npmi: 'NPMI',
-        fisher_p: 'Fisher p-value',
-        fdr_bh: 'FDR (BH q-value)',
-        oldest_pub_date: 'Oldest Publication',
-        source: 'Source',
-        pmids: 'PMIDs',
-      },
-    });
-    makeToast('Excel file downloaded', 'Success', 'success');
-  } catch (_error) {
-    makeToast('Export failed', 'Error', 'danger');
-  }
-};
-
-// Merge server fspec changes into local fields - preserve actions column
-const mergeFields = (inboundFields: FieldDefinition[]): FieldDefinition[] => {
-  const merged = inboundFields.map((f) => {
-    const existing = fields.value.find((x) => x.key === f.key);
-    return {
-      ...f,
-      filterable: existing?.filterable ?? false,
-      class: existing?.class ?? 'text-start',
-    };
-  });
-
-  // Always add actions column at the end for expand functionality
-  merged.push({
-    key: 'actions',
-    label: '',
-    sortable: false,
-    class: 'text-center',
-    filterable: false,
-  });
-
-  return merged;
-};
-
-// Cleanup: abort all in-flight publication requests on unmount
-onUnmounted(() => {
-  cancelAllPublicationFetches();
-});
-
-// Lifecycle
-onMounted(() => {
-  // Transform sort string into sortBy and sortDesc
-  const sortObject = sortStringToVariables(props.sortInput);
-  sortBy.value = sortObject.sortBy;
-  sort.value = props.sortInput;
-
-  // If we have a pre-loaded filter string, parse it
-  if (props.filterInput && props.filterInput !== 'null' && props.filterInput !== '') {
-    filter.value = filterStrToObj(props.filterInput, filter.value);
-  }
-
-  // Slight delay, then show table
-  setTimeout(() => {
-    loading.value = false;
-  }, 300);
-
-  // Load initial data
-  loadData();
-});
+  parsePmids,
+  truncateText,
+  getCompactTooltipText,
+} = usePubtatorGenesTable(props, emit);
 </script>
 
 <style scoped>
